refactor: move request handlers to `api/services` package
