+++ internals/api/services/benchmark.go
@@ -1,0 +1,10 @@
+package services
+
+import (
+	"log"
+	"net/http"
+)
+
+func BenchmarkHandler(w http.ResponseWriter, r *http.Request) {
+	log.Println("Benchmark route accessed")
+}
