+++ internals/api/services/helpers.go
@@ -1,4 +1,4 @@
-package api
+package services
 
 import (
 	"fmt"
